+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.17.22
+
+* added empty context manager